split into 3 files, some doc improvements
