+++ README.md
@@ -1,3 +1,3 @@
 # Jorropo's channel wrapper
 
-This is a wrapper around channels, it use usefull because it adds an error field which indicate how the request terminated, this can be red by multiple consumers.
+This is a wrapper around channels, it's usefull because it adds an error field which indicate how the request terminated, this can be red by multiple consumers.
